requirements.txt
transformers
torch
pillow
matplotlib
ipywidgets
streamlit
numpy
jupyter
ipython
